+++ packages/test/src/workflows/type-info/interceptors.ts
@@ -3,6 +3,16 @@ import { workflowInfo } from '@temporalio/workflow';
 import { workflowTypeInfo } from './models';
 
 export const interceptors = (): WorkflowInterceptors => ({
+  inbound: [
+    {
+      handleQuery(input, next) {
+        if (workflowInfo().workflowType !== 'queryTarget' || input.queryName !== 'order-alias') {
+          return next(input);
+        }
+        return next({ ...input, queryName: 'order' });
+      },
+    },
+  ],
   outbound: [
     {
       continueAsNew(input, next) {