+++ packages/workflow/src/internals.ts
@@ -889,15 +889,20 @@ export class Activator implements ActivationHandler {
       queryType === STACK_TRACE_QUERY_NAME ||
       queryType === ENHANCED_STACK_TRACE_QUERY_NAME;
     const interceptors = isInternalQuery ? [] : this.interceptors.inbound;
-    const execute = composeInterceptors(interceptors, 'handleQuery', this.queryWorkflowNextHandler.bind(this));
     const context = this.workflowSerializationContext();
+    const typeInfo = this.queryHandlers.get(queryType)?.typeInfo;
+    let outputTypeInfo = typeInfo?.outputType;
+    const execute = composeInterceptors(interceptors, 'handleQuery', (input) => {
+      outputTypeInfo = this.queryHandlers.get(input.queryName)?.typeInfo?.outputType;
+      return this.queryWorkflowNextHandler(input);
+    });
     execute({
       queryName: queryType,
-      args: arrayFromPayloads(this.payloadConverter, activation.arguments, context),
+      args: arrayFromPayloads(this.payloadConverter, activation.arguments, context, typeInfo?.inputTypes),
       queryId,
       headers: headers ?? {},
     }).then(
-      (result) => this.completeQuery(queryId, result),
+      (result) => this.completeQuery(queryId, result, outputTypeInfo),
       (reason) => this.failQuery(queryId, reason)
     );
   }
@@ -1390,10 +1395,13 @@ export class Activator implements ActivationHandler {
     if (this.workflowTaskError) throw this.workflowTaskError;
   }
 
-  private completeQuery(queryId: string, result: unknown): void {
+  private completeQuery(queryId: string, result: unknown, typeInfo?: TypeInfo): void {
     const context = this.workflowSerializationContext();
     this.pushCommand({
-      respondToQuery: { queryId, succeeded: { response: this.payloadConverter.toPayload(result, context) } },
+      respondToQuery: {
+        queryId,
+        succeeded: { response: toPayloadWithTypeInfo(this.payloadConverter, result, context, typeInfo) },
+      },
     });
   }